Initial working draft of the library.
Added a rough working draft of the library. The library functions by
authenticating the user using Puppeteer and saving the session cookies
to disk. The session cookies are then used to get the JSON web
token (JWT) associated with the authenticated user. This JWT is used as
as the bearer token in subsequent API requests.