Resolved an issue where the user was incorrectly assumed to be authenticated.
It was incorrectly assumed that expiration of the user's session would
be provided in the session cookies. The authentication logic now tries
to get the JWT/bearer token to determine if the session is active. If
the operation succeeds then the session is active; if it fails then
the user must re-authenticate.